# docs-site/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
...
  "info": {
    "contact": {
      "name": "tscrond",
      "url": "https://github.com/tscrond"
    },
    "description": "Fluxsend API endpoints",
    "license": {
      "name": "MIT",
      "url": "https://mit-license.org/"
    },
    "title": "FluxSend Developer API",
    "version": "1.7.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/workspaces/members/role": {
      "patch": {
        "description": "Updates the role of a workspace member when the caller has permission to manage members.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "Member role change request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Change member role",
        "tags": [
          "Workspaces"
        ]
      }
    },
    "/api/workspaces/rename": {
      "patch": {
        "description": "Updates the name and optionally the slug of a workspace owned by the current user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "Workspace rename request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Rename workspace",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://fluxsend.win/"
    }
  ]
}